+++ src/Publisher/Publisher.php
@@ -5,9 +5,9 @@ namespace Convenia\Pigeon\Publisher;
 use Convenia\Pigeon\Drivers\DriverContract;
 use Illuminate\Foundation\Application;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Str;
 use PhpAmqpLib\Message\AMQPMessage;
 use PhpAmqpLib\Wire\AMQPTable;
-use Webpatser\Uuid\Uuid;
 
 class Publisher implements PublisherContract
 {
@@ -61,7 +61,7 @@ class Publisher implements PublisherContract
         return array_merge([
             'content_type' => 'application/json',
             'content_encoding' => 'utf8',
-            'correlation_id' => Uuid::generate()->string,
+            'correlation_id' => Str::uuid(),
             'expiration' => 60000000,
             'app_id' => $this->app['config']['app_name'],
             'application_headers' => new AMQPTable($this->getHeaders()),